健康打卡，物资库存，入库记录

+++ goods-business/src/main/java/com/goods/business/service/impl/CategoryServiceImpl.java
@@ -2,8 +2,11 @@ package com.goods.business.service.impl;
 
 import com.goods.business.converter.ProductCategoryTreeNodeVOConverter;
 import com.goods.business.mapper.ProductCategoryMapper;
+import com.goods.business.mapper.ProductMapper;
 import com.goods.business.service.CategoryService;
+import com.goods.common.model.business.Product;
 import com.goods.common.model.business.ProductCategory;
+import com.goods.common.response.ResponseBean;
 import com.goods.common.utils.CategoryTreeBuilder;
 import com.goods.common.utils.ListPageUtils;
 import com.goods.common.vo.business.ProductCategoryTreeNodeVO;
@@ -12,8 +15,10 @@ import com.goods.common.vo.system.PageVO;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import tk.mybatis.mapper.entity.Example;
 
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
 
 /**
@@ -30,6 +35,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private ProductCategoryTreeNodeVOConverter productCategoryTreeNodeVOConverter;
+
+    @Autowired
+    private ProductMapper productMapper;
 
     @Override
     public PageVO<ProductCategoryTreeNodeVO> categoryTree(Integer pageNum, Integer pageSize) {
@@ -84,7 +92,25 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void delete(Long categoryId) {
+    public ResponseBean delete(Long categoryId) {
+        HashMap<Object, Object> map = new HashMap<>();
+        Example example = new Example(ProductCategory.class);
+        Example.Criteria criteria = example.createCriteria();
+        criteria.andEqualTo("pid", categoryId);
+        List<ProductCategory> categories = productCategoryMapper.selectByExample(example);
+        if (categories.size() != 0) {
+            map.put("errorMsg", "不能删除父级分类!");
+            return ResponseBean.error(map);
+        }
+        Example e = new Example(Product.class);
+        Example.Criteria c = e.createCriteria();
+        c.andEqualTo("threeCategoryId", categoryId);
+        List<Product> productList = productMapper.selectByExample(e);
+        if (productList.size() != 0) {
+            map.put("errorMsg", "分类下有物资引用!");
+            return ResponseBean.error(map);
+        }
         productCategoryMapper.deleteByPrimaryKey(categoryId);
+        return ResponseBean.success();
     }
 }
